+++ Engine/Code/Engine/Renderer/Texture1D.cpp
@@ -13,7 +13,7 @@ Texture1D::Texture1D(const RHIDevice* device, ID3D11Texture1D* dxTexture)
 
 void Texture1D::SetDebugName([[maybe_unused]] const std::string& name) const noexcept {
 #ifdef RENDER_DEBUG
-    _dx_tex->SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.data());
+    _dx_tex->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<unsigned int>(name.size()), name.data());
 #endif
 }
