+++ Engine/Code/Engine/Renderer/Texture3D.cpp
@@ -12,7 +12,7 @@ Texture3D::Texture3D(const RHIDevice* device, ID3D11Texture3D* dxTexture)
 
 void Texture3D::SetDebugName([[maybe_unused]] const std::string& name) const noexcept {
 #ifdef RENDER_DEBUG
-    _dx_tex->SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.data());
+    _dx_tex->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<unsigned int>(name.size()), name.data());
 #endif
 }
